+++ meinberlin/apps/documents/templates/meinberlin_documents/react_documents.html
@@ -3,4 +3,10 @@
 {{ ckeditor_media.js }}
 
 
-<div data-mb-widget="document-management" data-module="{{ module }}" data-chapters="{{ chapters }}" data-config="{{ config }}" data-reloadOnSuccess="{{ reload_on_success }}"></div>
+<div
+  data-mb-widget="document-management"
+  data-module="{{ module }}"
+  data-chapters="{{ chapters }}"
+  data-config="{{ config }}"
+  data-reloadOnSuccess="{{ reload_on_success }}">
+</div>
