--- meinberlin/apps/livequestions/assets/react_questions_present.jsx
@@ -1,13 +1,0 @@
-import PresentBox from './PresentBox'
-import React from 'react'
-import { createRoot } from 'react-dom/client'
-
-export function renderData (el) {
-  const props = JSON.parse(el.getAttribute('data-attributes'))
-  const root = createRoot(el)
-  root.render(
-    <React.StrictMode>
-      <PresentBox {...props} />
-    </React.StrictMode>
-  )
-}
