+++ meinberlin/assets/js/app.js
@@ -23,8 +23,6 @@ import {
 } from 'adhocracy4'
 
 import * as ReactMapTeaser from '../../apps/plans/assets/react_map_teaser.jsx'
-import * as ReactQuestions from '../../apps/livequestions/assets/react_questions.jsx'
-import * as ReactQuestionsPresent from '../../apps/livequestions/assets/react_questions_present.jsx'
 
 import * as Tether from 'tether'
 
@@ -48,9 +46,6 @@ function init () {
   ReactWidget.initialise('a4', 'reports', ReactReports.renderReports)
 
   ReactWidget.initialise('mb', 'mapTeaser', ReactMapTeaser.renderFilter)
-
-  ReactWidget.initialise('ie', 'questions', ReactQuestions.renderQuestions)
-  ReactWidget.initialise('ie', 'present', ReactQuestionsPresent.renderData)
 
   $('.timeline-carousel__item').slick({
     initialSlide: parseInt($('#timeline-carousel').attr('data-initial-slide')),
